Added a rough text rendering system.
Added Font class to support setup of freetype fonts and retaining necessary font info.
Text can be added into a layer/group simply by adding a Label object. Same thing as adding Sprite objects.
The constructor for this class currently resides in the layer class as there is an issue in calling the constructor in the Renderer class. This should be looked in to in the future.