Fixed camera shearing issue, added on scroll event
Camera shearing was fixed by resetting the aspect ratio everytime the window size is changed.
This makes it such that if the height is changed, the window contents will scale accordingly. And if width is changed, more stuff can be displayed on screen (think ultrawide monitor).

Added MouseWheel Scroll event