Redesigned adding colliders a little
Added a vector of shared ptr<colliders> in entities.
The Physics manager has a shared ptr vector to entities with colliders.
The physics manager update loop loops through each collider in each entity and performs collider checks.
Not very efficient but gets the job done.
To find a way to optimize this horrible nested loop