Test with 1k+ objects
This test allowed me to achieve the 0.14 fps dream.

The engine is inefficient (as expected). It runs extremely poorly with many objects.
The engine runs at a nice 300+fps at 144 objects. So i think try to keep it under 200 objects.

Also, try not to have too many colliders. The collision system is EXTREMELY inefficient.